Validating data before storing

+++ results.php
@@ -1,5 +1,6 @@
 <?php
   $file_name = "results.json";
+  header('Content-Type: application/json');
   if ($_SERVER['REQUEST_METHOD'] === 'GET')
     handleGETRequest($file_name);
   if ($_SERVER['REQUEST_METHOD'] === 'POST')
@@ -7,15 +8,17 @@
 
   function handleGETRequest($file_name) {
     $file_content = file_get_contents($file_name);
-    header('Content-Type: application/json');
     echo $file_content;
   }
 
   function handlePOSTRequest($file_name, $post) {
     $file_content = file_get_contents($file_name);
     $json = json_decode($file_content, true);
-    array_push($json, $post);
-    file_put_contents($file_name, json_encode($json));
+    if(array_key_exists("user", $post) && array_key_exists("result", $post) && is_numeric($post["result"])) {
+      $post["result"] = intval($post["result"]);
+      array_push($json, $post);
+      file_put_contents($file_name, json_encode($json));
+    }
     echo file_get_contents($file_name);
   }
 ?>
